Made tools work better offline
When offline, feather icons and font is not available.
This commit adds text defaults for when logos are not replaced.
Text is made italic on hover, to prevent width changes.
Tested with bold or font size increase, but it didn't look good.
Not the ideal solution, but good enough, mainly wanted the app
to be usable offline.

Signed-off-by: Ole Herman Schumacher Elgesem <oleherman93@gmail.com>