Added ability to draw multiple segments using wire tool
Wires are represented using SVG polyline and polygon.
If you "close" the drawing, by ending where you begin,
it's converted to a polygon.

There is no longer a global "tmp" wire. Instead
temporary objects, like the wire or dot you are currently
placing are tracked inside the SVG, using classes, similar
to how deletion works.

Items which will be deleted have delete class and are red.
Items which will be added have tmp class and are blue.

This commit does remove some functionality, like the protection
from placing the same dot twice, but this can be added back again
in a later commit. (With similar functionality for all shapes.)

Signed-off-by: Ole Herman Schumacher Elgesem <oleherman93@gmail.com>